[expo-notifications] Extract ExpoNotificationBehavior
